\name{annulus}
\alias{annulus}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Annulus toy data set}
\description{
  The annulus data frame has 1000 rows and 2 columns.  500 data points
  are from the noisy annulus and 500 data points reside within the annulus.
}
\usage{
annulus
}
%- maybe also 'usage' for other objects documented here.
\format{
  Data are in two dimensions.
}
\keyword{ data }% __ONLY ONE__ keyword per line
