\name{Chainlink}
\alias{Chainlink}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Chainlink toy clustering data set}
\description{
  The Chainlink data frame has 1000 rows and 3 columns.  The data are of
  two interlocking 3-dimensional rings. 500 data points are from one
  ring and 500 from the other ring.
}
\usage{
Chainlink
}
%- maybe also 'usage' for other objects documented here.
\format{
  The data are in 3 dimensions, C1, C2, and C3.
}
\source{
  \url{http://www.uni-marburg.de/fb12/datenbionik/data?language_sync=1}
   }
\references{
  Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on Self-Organizing Maps, Paris, France, (2005) , pp. 75-82  }
\keyword{ data }% __ONLY ONE__ keyword per line
