% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annulus}
\alias{annulus}
\title{Annulus toy data set}
\format{Data are in two dimensions.}
\usage{
annulus
}
\description{
The annulus data frame has 1000 rows and 2 columns. 500 data points are
from the noisy annulus and 500 data points reside within the annulus.
}
\keyword{data}
