% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Chainlink}
\alias{Chainlink}
\title{Chainlink toy clustering data set}
\format{The data are in 3 dimensions, C1, C2, and C3.}
\source{
\url{http://www.uni-marburg.de/fb12/datenbionik/data?language_sync=1}
}
\usage{
Chainlink
}
\description{
The Chainlink data frame has 1000 rows and 3 columns.  The data are of two
interlocking 3-dimensional rings. 500 data points are from one ring and 500
from the other ring.
}
\references{
Ultsch, A.: Clustering with SOM: U*C, In Proc. Workshop on
Self-Organizing Maps, Paris, France, (2005) , pp. 75-82
}
\keyword{data}
